requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
PyNaCl==1.3.0
async-timeout==3.0.1
datetime
redis
python-time
colorama 
random2
requests
nekos.py
urllib3
math22
bs4 
praw
python-dotenv
asyncio
aiohttp
simpy.io
urllib3
requests
discord 
discord.py
game
js
typing
numpy  
operator-courier
utils
httpx
traceback2
youtube_dl
discord.py[voice]
opus-api
ffmpeg
apscheduler
python-settings
mediawikiapi
lyricsgenius
umongo
mongo


